+++ insider/src/main/java/com/kh/insider/controller/MemberController.java
@@ -22,8 +22,12 @@ import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
 import com.kh.insider.dto.BoardDto;
+import com.kh.insider.dto.FollowDto;
+import com.kh.insider.dto.FollowWithProfileDto;
+import com.kh.insider.dto.FollowerWithProfileDto;
 import com.kh.insider.dto.MemberDto;
 import com.kh.insider.repo.BoardRepo;
+import com.kh.insider.repo.FollowRepo;
 import com.kh.insider.repo.MemberRepo;
 import com.kh.insider.service.MemberService;
 import com.kh.insider.service.SocialLoginService;
@@ -42,279 +46,177 @@ import lombok.extern.slf4j.Slf4j;
 @CrossOrigin
 @RequestMapping("/member")
 public class MemberController {
-	
-	@Value("${cos.key}")
-	private String cosKey;
-	
-	
-	@Autowired
-	private MemberRepo memberRepo;
-	@Autowired
-	private MemberService memberService;
-	@Autowired
-	private SettingRepo settingRepo;
-	
-	@Autowired
-	private SocialLoginService socialLoginService;
-	
-	@Autowired
-	private BoardRepo boardRepo;
-	
-	@GetMapping("/join")
-	public String join() {
-		return "member/join";
-	}
-	
-	@PostMapping("/join")
-	public String join(@ModelAttribute MemberDto dto) {
-		memberRepo.join(dto);
-		//기본 회원설정값 생성(추후 수정 필요)
-//		settingRepo.basicInsert(dto.getMemberNo());
-		return "redirect:/";
-	}
-	
-	@GetMapping("/login")
-	public String login() {
-		return "member/login";
-	}
-	
-	@PostMapping("/login")
-	public String login(HttpSession session, @ModelAttribute MemberDto dto, RedirectAttributes attr) {
-		
-	MemberDto findMember = memberRepo.login(dto.getMemberEmail(), dto.getMemberPassword());
-	
-	if(findMember == null) {
-		int result = 0;
-		attr.addFlashAttribute("result",result);
-		return "redirect:login";
-	}
-	memberRepo.updateLoginTime(findMember.getMemberNo());
-	session.setAttribute("memberNo",findMember.getMemberNo());
-	session.setAttribute("socialUser", findMember);
-	session.setAttribute("memberNick", findMember.getMemberNick());
-		
-	return "redirect:/";
-	}
-	
-	@GetMapping("/logout")
-	public String logout(HttpSession session) {
-		session.removeAttribute("memberNo");
-		session.removeAttribute("socialUser");
-		session.removeAttribute("member");
-		session.removeAttribute("memberNick");
-		
-		return "redirect:/";
-	}
-	
-	@GetMapping("/{memberNick}")
-	public String myPage(@PathVariable String memberNick, Model model, HttpSession session) {
-		
-		// 프로필 정보 불러오기
-		MemberDto findMember = memberRepo.findByNickName(memberNick);
-		// 로그인한 사용자
-		MemberDto loginUser = (MemberDto)session.getAttribute("socialUser");
-		// 본인 프로필 인지 여부
-		boolean isOwner = loginUser.getMemberNick().equals(memberNick);
-		// 전체 게시물 개수
-		int totalPostCount = boardRepo.getTotalPostCount(findMember.getMemberNo());
-		
-		// 마이페이지 게시물 조회
-		List<BoardDto> getTotalMyPost = boardRepo.getTotalMyPost(findMember.getMemberNo());
-		
-		model.addAttribute("getTotalMyPost",getTotalMyPost);
-		model.addAttribute("totalPostCount",totalPostCount);
-		model.addAttribute("memberDto",findMember);
-		model.addAttribute("isOwner",isOwner);
-		
-		
-		
-		return "/member/mypage";
-	}
-	
-	@GetMapping("/emailCheck")
-	@ResponseBody
-	public String isEmailDuplicated(@RequestParam String memberEmail) throws Exception {
-		
-		int result = memberRepo.isEmailDuplicated(memberEmail);
-		log.debug("result = {}",result);
-		
-		if(result != 0) {
-			return "fail";
-		}else {
-			return "success";
-		}
-	}
-	
-	@GetMapping("/sendMail")
-	@ResponseBody
-	public String sendMail(@RequestParam String memberEmail) {
-		
-		int num = memberService.sendEmail(memberEmail);
-		System.out.println(Integer.toString(num));
-		return Integer.toString(num);
-		
-	}
-	
-	
-	
-	@GetMapping("/nickCheck")
-	@ResponseBody
-	public String isNickDuplicated(@RequestParam String memberNick) throws Exception{
-		
-		int result = memberRepo.isNickDuplicated(memberNick);
-		if(result != 0) {
-			return "fail";
-		}else {
-			return "success";
-		}
-	}
-	
-	@GetMapping("/passwordChange")
-	public String passwordChange() {
-		return "member/passwordChange";
-	}
-	
-	@PostMapping("/passwordChange")
-	@ResponseBody
-	public String passwordChange(@RequestBody MemberDto dto) {
-		// 임시 비밀번호
-		String generatTempPassword = memberService.generatTempPassword();
-		dto.setMemberPassword(generatTempPassword);
-		memberRepo.updateTempPassword(dto);
-		
-		return generatTempPassword;
-	}
-	
-	
-	
-	// 카카오 로그인
-	@GetMapping("/auth/kakao/callback")
-	public String kakaoCallback(String code,
-			HttpSession session, KakaoResponseVO token, Model model) throws URISyntaxException { // Data를 리턴해주는 컨트롤러 함수
-				
-		token = socialLoginService.kakaoTokenCreate(code);
-		MemberDto kakaoUser = new MemberDto();
-		KakaoProfileVO profile = socialLoginService.kakaoLogin(code,token);
-		Long memberNo = profile.getId();
-		String memberEmail = profile.kakao_account.getEmail();
-		String memberPw = cosKey;
-		MemberDto originalMember = memberRepo.findByEmail(profile.kakao_account.getEmail());
-		
-		
-		
-		if(originalMember == null) {
-			System.out.println("회원가입을 진행합니다..");
-			kakaoUser.setMemberNo(memberNo);
-			kakaoUser.setMemberEmail(memberEmail);
-			kakaoUser.setMemberPassword(memberPw);
-				
-		}else {
-			System.out.println("기존회원이므로 로그인을 진행합니다.");
-			// 로그인 시각 갱신
-			memberRepo.updateLoginTime(originalMember.getMemberNo());
-			// 회원정보
-			session.setAttribute("socialUser",originalMember);
-			session.setAttribute("memberNo", originalMember.getMemberNo());
-			// 토큰정보
-			session.setAttribute("member",token.getAccess_token());
-			// 리프레시 토큰 정보
-			session.setAttribute("refresh_token",token.getRefresh_token());
-			return "redirect:/";
-		}
-		// addInfo로 넘길 정보
-		session.setAttribute("loginUser",kakaoUser);
-		session.setAttribute("member",token.getAccess_token());
-		session.setAttribute("refresh_token",token.getRefresh_token());
-		
-		
-		return "redirect:/member/addInfo";
-	}
-	
-	@GetMapping("/addInfo")
-	public String addInfo(Model model, HttpSession session) {
-
-		MemberDto loginUser =(MemberDto) session.getAttribute("loginUser");
-
-		model.addAttribute("loginUser",loginUser);
-		
-		return "member/addInfo";
-	}
-	
-	@PostMapping("/addInfo")
-	public String addInfo(@ModelAttribute MemberDto dto) {
-	
-		memberRepo.socialJoin(dto);
-		
-		return "redirect:/";
-	}
-	
-	@GetMapping("/login/oauth_google_check")
-	public String googleCallback(String code, GoogleResponseVO response, HttpSession session) throws URISyntaxException {
-		response =  socialLoginService.googleTokenCreate(code);
-		MemberDto googleUser = new MemberDto();
-		GoogleProfileVO profile = socialLoginService.googleLogin(code,response);
-		MemberDto originalMember = memberRepo.findByEmail(profile.getEmail());
-		UUID uuid = UUID.randomUUID();
-		String uuidString = uuid.toString();
-		  // 하이픈(-) 제거
-        String uuidWithoutHyphens = uuidString.replaceAll("-", "");
-        // 숫자 부분 추출
-        String numbersOnly = uuidWithoutHyphens.replaceAll("\\D", "");
-        numbersOnly = numbersOnly.substring(0,10);
-		
-		Long memberNo = Long.parseLong(numbersOnly);
-		String memberEmail = profile.getEmail();
-		String memberPw = cosKey;
-		String memberName = profile.getName();
-		
-		
-		
-		if(originalMember == null) {
-			System.out.println("회원가입을 진행합니다");
-			googleUser.setMemberNo(memberNo);
-			googleUser.setMemberEmail(memberEmail);
-			googleUser.setMemberName(memberName);
-			googleUser.setMemberPassword(memberPw);
-
-			
-		}else {
-			System.out.println("기존회원이므로 로그인을 진행합니다.");
-			// 회원정보
-			// 로그인 시각 갱신
-			memberRepo.updateLoginTime(googleUser.getMemberNo());
-			session.setAttribute("socialUser",originalMember);
-			session.setAttribute("memberNo", originalMember.getMemberNo());
-			// 토큰정보
-			session.setAttribute("member",response.getAccess_token());
-			session.setAttribute("refresh_token",response.getRefresh_token());
-			return "redirect:/";
-		}
-		// addInfo로 넘길 정보
-		session.setAttribute("loginUser",googleUser);
-		session.setAttribute("member",response.getAccess_token());
-		session.setAttribute("refresh_token",response.getRefresh_token());
-		
-
-		return "redirect:/member/addInfo";
-	}
-	
-	@GetMapping("/facebook/auth")
-	@ResponseBody
-	public String facebookLogin(String code, FacebookResponseVO response) throws URISyntaxException{
-		
-		response = socialLoginService.facebookTokenCreate(code);
-		FacebookProfileVO profile = socialLoginService.facebookLogin(code, response);
-		System.out.println(profile);
-		
-		return profile.toString();
-	}
-	
-	
-	
-//	환경설정 페이지
-	@GetMapping("/setting")
-	public String setting() {
-		return"member/setting";
-	}
+   
+   @Value("${cos.key}")
+   private String cosKey;
+   
+   @Autowired
+   private MemberRepo memberRepo;
+   @Autowired
+   private MemberService memberService;
+   @Autowired
+   private SettingRepo settingRepo;
+   
+   @Autowired
+   private SocialLoginService socialLoginService;
+   
+   @Autowired
+   private BoardRepo boardRepo;
+   
+   @Autowired
+   private FollowRepo followRepo;
+   
+   @GetMapping("/join")
+   public String join() {
+      return "member/join";
+   }
+   
+   @PostMapping("/join")
+   public String join(@ModelAttribute MemberDto dto) {
+      memberRepo.join(dto);
+      //기본 회원설정값 생성(추후 수정 필요)
+//      settingRepo.basicInsert(dto.getMemberNo());
+      return "redirect:/";
+   }
+   
+   @GetMapping("/login")
+   public String login() {
+      return "member/login";
+   }
+   
+   @PostMapping("/login")
+   public String login(HttpSession session, @ModelAttribute MemberDto dto, RedirectAttributes attr) {
+      
+   MemberDto findMember = memberRepo.login(dto.getMemberEmail(), dto.getMemberPassword());
+   
+   if(findMember == null) {
+      int result = 0;
+      attr.addFlashAttribute("result",result);
+      return "redirect:login";
+   }
+   memberRepo.updateLoginTime(findMember.getMemberNo());
+   session.setAttribute("memberNo",findMember.getMemberNo());
+   session.setAttribute("socialUser", findMember);
+   session.setAttribute("memberNick", findMember.getMemberNick());
+      
+   return "redirect:/";
+   }
+   
+   @GetMapping("/logout")
+   public String logout(HttpSession session) {
+      session.removeAttribute("memberNo");
+      session.removeAttribute("socialUser");
+      session.removeAttribute("member");
+      session.removeAttribute("memberNick");
+      
+      return "redirect:/";
+   }
+   
+   @GetMapping("/{memberNick}")
+   public String myPage(@PathVariable String memberNick, Model model, HttpSession session) {
+      
+      // 프로필 정보 불러오기
+      MemberDto findMember = memberRepo.findByNickName(memberNick);
+      // 로그인한 사용자
+      MemberDto loginUser = (MemberDto)session.getAttribute("socialUser");
+      // 본인 프로필 인지 여부
+      boolean isOwner = loginUser.getMemberNick().equals(memberNick);
+      // 전체 게시물 개수
+      int totalPostCount = boardRepo.getTotalPostCount(findMember.getMemberNo());  
+      // 본인 팔로우 개수 
+      int totalFollowCount = followRepo.getFollowNumber(findMember.getMemberNo());
+      // 본인 팔로워 개수
+      int totalFollowerCount = followRepo.getFollowerNumber(findMember.getMemberNo());
+      
+      // 팔로워 목록
+      List<FollowerWithProfileDto> followerList = followRepo.getFollowerList(findMember.getMemberNo());
+      
+      // 팔로우 목록
+      List<FollowWithProfileDto> followList = followRepo.getFollowList(findMember.getMemberNo());
+      
+      // 마이페이지 게시물 조회
+      List<BoardDto> getTotalMyPost = boardRepo.getTotalMyPost(findMember.getMemberNo());
+      
+      model.addAttribute("getTotalMyPost",getTotalMyPost);
+      model.addAttribute("totalPostCount",totalPostCount);
+      model.addAttribute("totalFollowCount",totalFollowCount);
+      model.addAttribute("totalFollowerCount",totalFollowerCount);
+      model.addAttribute("followerList",followerList);
+      model.addAttribute("followList",followList);
+      model.addAttribute("memberDto",findMember);
+      model.addAttribute("isOwner",isOwner);
+            
+      return "/member/mypage";
+   }
+   
+   @GetMapping("/emailCheck")
+   @ResponseBody
+   public String isEmailDuplicated(@RequestParam String memberEmail) throws Exception {
+      
+      int result = memberRepo.isEmailDuplicated(memberEmail);
+      log.debug("result = {}",result);
+      
+      if(result != 0) {
+         return "fail";
+      }else {
+         return "success";
+      }
+   }
+   
+   @GetMapping("/sendMail")
+   @ResponseBody
+   public String sendMail(@RequestParam String memberEmail) {
+      
+      int num = memberService.sendEmail(memberEmail);
+      System.out.println(Integer.toString(num));
+      return Integer.toString(num);
+      
+   }
+   
+   @GetMapping("/nickCheck")
+   @ResponseBody
+   public String isNickDuplicated(@RequestParam String memberNick) throws Exception{
+      
+      int result = memberRepo.isNickDuplicated(memberNick);
+      if(result != 0) {
+         return "fail";
+      }else {
+         return "success";
+      }
+   }
+   
+   @GetMapping("/passwordChange")
+   public String passwordChange() {
+      return "member/passwordChange";
+   }
+   
+   @PostMapping("/passwordChange")
+   @ResponseBody
+   public String passwordChange(@RequestBody MemberDto dto) {
+      // 임시 비밀번호
+      String generatTempPassword = memberService.generatTempPassword();
+      dto.setMemberPassword(generatTempPassword);
+      memberRepo.updateTempPassword(dto);
+      
+      return generatTempPassword;
+   }
+   
+   
+   @GetMapping("/facebook/auth")
+   @ResponseBody
+   public String facebookLogin(String code, FacebookResponseVO response) throws URISyntaxException{
+      
+      response = socialLoginService.facebookTokenCreate(code);
+      FacebookProfileVO profile = socialLoginService.facebookLogin(code, response);
+      System.out.println(profile);
+      
+      return profile.toString();
+   }
+   
+//   환경설정 페이지
+   @GetMapping("/setting")
+   public String setting() {
+      return"member/setting";
+   }
 
 }
