+++ insider/src/main/java/com/kh/insider/dto/FollowerWithProfileDto.java
@@ -10,7 +10,7 @@ public class FollowerWithProfileDto {
 	private long followFollower;
 	private Date followTime;
 	private int followAllow;
-	private long memberName;
+	private String memberName;
 	private String memberNick;
 	private int attachmentNo;
 }
