+++ insider/src/main/java/com/kh/insider/service/MemberServiceImpl.java
@@ -17,74 +17,74 @@ import org.springframework.util.StringUtils;
 
 @Repository
 public class MemberServiceImpl implements MemberService{
-	
-	private static final String CHARACTERS = "ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789!@#$%^&*()-_=+";
-	
-	@Autowired
-	private JavaMailSender sender;
-	Random random = new Random();
+   
+   private static final String CHARACTERS = "ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789!@#$%^&*()-_=+";
+   
+   @Autowired
+   private JavaMailSender sender;
+   Random random = new Random();
 
-	@Override
-	public int RandomCode() {
-		
-		
-		
-		return random.nextInt(888888)+111111;
-	}
+   @Override
+   public int RandomCode() {
+      
+      
+      
+      return random.nextInt(888888)+111111;
+   }
 
-	public int sendEmail(String email) {
-	     int num = RandomCode();
+   public int sendEmail(String email) {
+        int num = RandomCode();
 
-	    /* 이메일 전송 */
-	    MimeMessage message = sender.createMimeMessage();
-	    try {
-	        MimeMessageHelper helper = new MimeMessageHelper(message, false, "UTF-8");
-	        helper.setTo(email);
-	        helper.setSubject("Insider 회원가입 인증번호 입니다");
+       /* 이메일 전송 */
+       MimeMessage message = sender.createMimeMessage();
+       try {
+           MimeMessageHelper helper = new MimeMessageHelper(message, false, "UTF-8");
+           helper.setTo(email);
+           helper.setSubject("Insider 회원가입 인증번호 입니다");
 
-	        ClassPathResource resource = new ClassPathResource("templates/email1.jsp");
+           ClassPathResource resource = new ClassPathResource("templates/email1.jsp");
 
-	        Scanner sc = null;
-	        try {
-	            sc = new Scanner(resource.getFile());
-	        } catch (IOException e) {
-	            e.printStackTrace();
-	        }
+           Scanner sc = null;
+           try {
+               sc = new Scanner(resource.getFile());
+           } catch (IOException e) {
+               e.printStackTrace();
+           }
 
-	        StringBuilder emailContent = new StringBuilder(); // StringBuilder를 사용하여 문자열 조합
+           StringBuilder emailContent = new StringBuilder(); // StringBuilder를 사용하여 문자열 조합
 
-	        while (sc.hasNextLine()) {
-	            emailContent.append(sc.nextLine());
-	        }
+           while (sc.hasNextLine()) {
+               emailContent.append(sc.nextLine());
+           }
 
-	        // 인증번호 값으로 문자열 치환
-	        emailContent.replace(emailContent.indexOf("${num}"), emailContent.indexOf("${num}") + "${num}".length(), String.valueOf(num));
+           // 인증번호 값으로 문자열 치환
+           emailContent.replace(emailContent.indexOf("${num}"), emailContent.indexOf("${num}") + "${num}".length(), String.valueOf(num));
 
-	        helper.setText(emailContent.toString(), true);
+           helper.setText(emailContent.toString(), true);
 
-	        sender.send(message);
-	    } catch (MessagingException e) {
-	        e.printStackTrace();
-	    }
-	    
-	    return num;
-	}
+           sender.send(message);
+       } catch (MessagingException e) {
+           e.printStackTrace();
+       }
+       
+       return num;
+   }
 
-	@Override
-	public String generatTempPassword() {
-		
-		
-		StringBuilder tempPassword = new StringBuilder();
-		
-		for(int i = 0; i<10; i++) {
-			int index = random.nextInt(CHARACTERS.length());
-			char randomChar = CHARACTERS.charAt(index);
-			tempPassword.append(randomChar);
-		}
-		
-		return tempPassword.toString();
-		
-	}
-	
+   @Override
+   public String generatTempPassword() {
+      
+      
+      StringBuilder tempPassword = new StringBuilder();
+      
+      for(int i = 0; i<10; i++) {
+         int index = random.nextInt(CHARACTERS.length());
+         char randomChar = CHARACTERS.charAt(index);
+         tempPassword.append(randomChar);
+      }
+      
+      return tempPassword.toString();
+      
+   }
+   
 
 }
