+++ insider/src/main/resources/application.properties
@@ -58,9 +58,9 @@ custom.email.password=murswhqmwpixibld
 
 
 # kakao login
-kakao.clientId =1b308937b1aec37f7b4bc57faeb4931b
+kakao.clientId =f32aed858ba7f7223af2f0954ddd3854
 kakao.grantType = authorization_code
-kakao.redirectUri = http://localhost:8080/member/auth/kakao/callback
+kakao.redirectUri = http://localhost:8080/kakao/login
 
 # facebook login
 facebook.client-id=1721778684918582
@@ -70,7 +70,7 @@ facebook.client-secret=769adcd24b74fdecbdd86c90f00fcd41
 # google login
 google.client-id=197694978566-bljc0eo7lnf071parv36ntrenp3g69eb.apps.googleusercontent.com
 google.client-secret=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
-google.redirect-uri=http://localhost:8080/member/login/oauth_google_check
+google.redirect-uri=http://localhost:8080/google/login
 google.grant-type=authorization_code
 
 
@@ -84,4 +84,3 @@ cos.key=cos1234
 #server.ssl.key-store-password=123456
 #server.ssl.key-store-type=PKCS12
 #server.ssl.key-alias=bns-ssl
-
