+++ insider/src/main/webapp/WEB-INF/views/member/login.jsp
@@ -8,21 +8,21 @@
 <title>Insert title here</title>
 <link rel="stylesheet" href="/static/css/commons.css">
 <!-- BootStrap CDN -->
-	<link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0-alpha3/dist/css/bootstrap.min.css" rel="stylesheet" type="text/css">
-	<style>
-	.logo{
-		font-size: 50px;
-		color: black;
-		text-decoration: none;
-	}
-	</style>
-</head>		
+   <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0-alpha3/dist/css/bootstrap.min.css" rel="stylesheet" type="text/css">
+   <style>
+   .logo{
+      font-size: 50px;
+      color: black;
+      text-decoration: none;
+   }
+   </style>
+</head>      
 <body>
 
-		        <div class="container col-lg-3 card p-5 mt-5" style="display:flex; justify-content: center" id="app">
-		        <div>
-		         <div class="text-center mb-3">         	
-		         	<a href="/" class="logo"><img src="/static/image/logo.png" style="width:50px; height:50px;" >insider</a>
+              <div class="container col-lg-3 card p-5 mt-5" style="display:flex; justify-content: center" id="app">
+              <div>
+               <div class="text-center mb-3">            
+                  <a href="/" class="logo"><img src="/static/image/logo.png" style="width:50px; height:50px;" >insider</a>
                 </div>
             <form action="login" method="post" @submit="isEmpty" >
                 <div class="mb-3 row">
@@ -33,31 +33,31 @@
                 </div>
                 <div class="mb-3 row">
                     <p v-if="showEmptyWarning" class="email-warning-message">아이디 혹은 비밀번호를 입력해주세요</p>
-                     <p v-if="${result == 0} && !showEmptyWarning">아이디 혹은 비밀번호를 일치하지 않습니다 </p>	
+                     <p v-if="${result == 0} && !showEmptyWarning">아이디 혹은 비밀번호를 일치하지 않습니다 </p>   
                 </div>
                 <div class="row mb-3">
                     <button class="btn btn-primary" type="submit">로그인</button>
                 </div> 
                   
             </form>          
-            	<div style="display:flex; justify-content: space-between; flex-direction: column; align-items: center;" >
-            	 <div class="row mb-3">
-            <a href="https://kauth.kakao.com/oauth/authorize?client_id=1b308937b1aec37f7b4bc57faeb4931b&redirect_uri=	
-			http://localhost:8080/member/auth/kakao/callback&
-			response_type=code&scope=account_email,gender,age_range&prompt=login">
-            	 <img src="/static/image/social/kakao_login.png">
+               <div style="display:flex; justify-content: space-between; flex-direction: column; align-items: center;" >
+                <div class="row mb-3">
+            <a href="https://kauth.kakao.com/oauth/authorize?client_id=f32aed858ba7f7223af2f0954ddd3854&redirect_uri=   
+         http://localhost:8080/kakao/login&
+         response_type=code&scope=account_email&prompt=login">
+                <img src="/static/image/social/kakao_login.png">
             </a>
             </div>
              <div class="row mb-3">
              <a href="https://accounts.google.com/o/oauth2/v2/auth?client_id=197694978566-bljc0eo7lnf071parv36ntrenp3g69eb.apps.googleusercontent.com&
-				redirect_uri=http://localhost:8080/member/login/oauth_google_check&response_type=code
-				&scope=email%20profile%20openid
-				&access_type=offline&prompt=login">
-            	 <img src="/static/image/social/google_login.png">
+            redirect_uri=http://localhost:8080/google/login&response_type=code
+            &scope=email%20profile%20openid
+            &access_type=offline&prompt=login">
+                <img src="/static/image/social/google_login.png">
             </a>
             </div>
              <div class="row text-center">
-                	<a class="mt-5" href="/member/passwordChange">비밀번호를 잊으셨나요?</a>
+                   <a class="mt-5" href="/member/passwordChange">비밀번호를 잊으셨나요?</a>
                 </div>         
          <!--  
             <div class="row mb-3">
@@ -70,14 +70,14 @@ redirect_uri=https://localhost:8080/member/facebook/auth&scope=public_profile,em
            </div>
         </div>
         <div class="container col-lg-3 card p-3 mt-3" style="display:flex;">
-        	<div class="text-center">
-        		계정이 없으신가요?
-        		<a href="join">가입하기</a>
-        	</div>
+           <div class="text-center">
+              계정이 없으신가요?
+              <a href="join">가입하기</a>
+           </div>
         </div>
 
         <div>
-	        소셜유저 : ${sessionScope.socialUser}, 멤버 : ${sessionScope.member} , 세션번호 : ${sessionScope.memberNo}			
+           소셜유저 : ${sessionScope.socialUser}, 멤버 : ${sessionScope.member} , 세션번호 : ${sessionScope.memberNo}         
         </div>
 
 
@@ -86,7 +86,7 @@ redirect_uri=https://localhost:8080/member/facebook/auth&scope=public_profile,em
 
         <script src="https://unpkg.com/vue@3.2.36"></script>
         <script>
-    	
+       
           const app = Vue.createApp({
                 // 데이터 설정 영역
                 data(){
@@ -108,7 +108,7 @@ redirect_uri=https://localhost:8080/member/facebook/auth&scope=public_profile,em
                             event.preventDefault();
                             this.showEmptyWarning = true;
                         }else{
-                        	this.showEmptyWarning = false;
+                           this.showEmptyWarning = false;
                         }
                     },
